requirements.txt
PySide6>=6.6.0
minimalmodbus>=2.1.1
pyserial>=3.5
pyqtgraph>=0.13.3
numpy>=1.24.0


